test(stories): move utility classes story files non-components to miscellaneous folder
